First draft of supporting LanguageTool for grammar and spell checks.
LanguageTool will be asked as an additional check if <LanguageToolUrl> is configured in the youshouldspellcheck.config.xml of the project. No CodeFix support yet. Only works for attributes at the moment. Probably needs bigger changes to the config file to control more precisely which analysis should use local dictionary or/and LanguageTool server.
(see Issue #5)